services/code-generation/core/deployment/kubernetes_manager.py
"""
Kubernetes infrastructure manager for deployment system
"""
import kubernetes
import logging
from typing import Dict, List, Any
from .base import CloudProviderManager, CloudProvider, InfrastructureResource, InfrastructureType

logger = logging.getLogger(__name__)

class KubernetesManager(CloudProviderManager):
    """Kubernetes infrastructure manager"""
    
    def __init__(self, config: Dict[str, Any]):
        super().__init__(CloudProvider.KUBERNETES, config)
        
        # Load kubeconfig
        if config.get("kubeconfig_path"):
            kubernetes.config.load_kube_config(config["kubeconfig_path"])
        else:
            kubernetes.config.load_incluster_config()
        
        self.v1 = kubernetes.client.CoreV1Api()
        self.apps_v1 = kubernetes.client.AppsV1Api()
        self.networking_v1 = kubernetes.client.NetworkingV1Api()
    
    async def create_infrastructure(
        self,
        resources: List[Dict[str, Any]],
        tags: Dict[str, str] = None
    ) -> List[InfrastructureResource]:
        """Create Kubernetes resources"""
        
        created_resources = []
        
        for resource_config in resources:
            resource_type = InfrastructureType(resource_config["type"])
            
            try:
                if resource_type == InfrastructureType.COMPUTE:
                    resource = await self._create_deployment(resource_config, tags)
                elif resource_type == InfrastructureType.LOAD_BALANCER:
                    resource = await self._create_service(resource_config, tags)
                elif resource_type == InfrastructureType.NETWORK:
                    resource = await self._create_ingress(resource_config, tags)
                else:
                    logger.warning(f"Unsupported Kubernetes resource type: {resource_type}")
                    continue
                
                created_resources.append(resource)
                
            except Exception as e:
                logger.error(f"Failed to create Kubernetes resource {resource_config['name']}: {e}")
                raise
        
        return created_resources
    
    async def destroy_infrastructure(self, resource_ids: List[str]) -> bool:
        """Destroy Kubernetes resources"""
        try:
            for resource_id in resource_ids:
                namespace, name = resource_id.split("/")
                # Implement destruction logic based on resource type
                pass
            return True
        except Exception as e:
            logger.error(f"Failed to destroy Kubernetes resources: {e}")
            return False
    
    async def get_resource_status(self, resource_id: str) -> str:
        """Get status of Kubernetes resource"""
        try:
            namespace, name = resource_id.split("/")
            # Implement status checking logic
            return "running"
        except Exception as e:
            logger.error(f"Failed to get Kubernetes resource status: {e}")
            return "unknown"
    
    async def scale_resource(self, resource_id: str, scale_config: Dict[str, Any]) -> bool:
        """Scale Kubernetes resource"""
        try:
            namespace, name = resource_id.split("/")
            # Implement scaling logic
            return True
        except Exception as e:
            logger.error(f"Failed to scale Kubernetes resource: {e}")
            return False
    
    async def _create_deployment(
        self,
        config: Dict[str, Any],
        tags: Dict[str, str] = None
    ) -> InfrastructureResource:
        """Create Kubernetes deployment"""
        
        name = config["name"]
        namespace = config.get("namespace", "default")
        
        labels = {"app": name}
        if tags:
            labels.update(tags)
        
        deployment = kubernetes.client.V1Deployment(
            metadata=kubernetes.client.V1ObjectMeta(
                name=name,
                namespace=namespace,
                labels=labels
            ),
            spec=kubernetes.client.V1DeploymentSpec(
                replicas=config.get("replicas", 1),
                selector=kubernetes.client.V1LabelSelector(
                    match_labels={"app": name}
                ),
                template=kubernetes.client.V1PodTemplateSpec(
                    metadata=kubernetes.client.V1ObjectMeta(
                        labels={"app": name}
                    ),
                    spec=kubernetes.client.V1PodSpec(
                        containers=[
                            kubernetes.client.V1Container(
                                name=name,
                                image=config.get("image"),
                                ports=[
                                    kubernetes.client.V1ContainerPort(
                                        container_port=config.get("port", 80)
                                    )
                                ],
                                env=[
                                    kubernetes.client.V1EnvVar(name=k, value=v)
                                    for k, v in config.get("environment", {}).items()
                                ],
                                resources=kubernetes.client.V1ResourceRequirements(
                                    requests={
                                        "cpu": config.get("cpu_request", "100m"),
                                        "memory": config.get("memory_request", "128Mi")
                                    },
                                    limits={
                                        "cpu": config.get("cpu_limit", "500m"),
                                        "memory": config.get("memory_limit", "512Mi")
                                    }
                                )
                            )
                        ]
                    )
                )
            )
        )
        
        response = self.apps_v1.create_namespaced_deployment(
            namespace=namespace,
            body=deployment
        )
        
        return InfrastructureResource(
            resource_id=f"{namespace}/{name}",
            name=name,
            resource_type=InfrastructureType.COMPUTE,
            provider=CloudProvider.KUBERNETES,
            region="",
            config=config,
            status="creating",
            metadata={
                "namespace": namespace,
                "replicas": config.get("replicas", 1),
                "image": config.get("image")
            }
        )
    
    async def _create_service(
        self,
        config: Dict[str, Any],
        tags: Dict[str, str] = None
    ) -> InfrastructureResource:
        """Create Kubernetes service"""
        
        name = config["name"]
        namespace = config.get("namespace", "default")
        
        labels = {"app": name}
        if tags:
            labels.update(tags)
        
        service = kubernetes.client.V1Service(
            metadata=kubernetes.client.V1ObjectMeta(
                name=name,
                namespace=namespace,
                labels=labels
            ),
            spec=kubernetes.client.V1ServiceSpec(
                selector={"app": config.get("selector", name)},
                ports=[
                    kubernetes.client.V1ServicePort(
                        port=config.get("port", 80),
                        target_port=config.get("target_port", 80),
                        protocol="TCP"
                    )
                ],
                type=config.get("service_type", "ClusterIP")
            )
        )
        
        response = self.v1.create_namespaced_service(
            namespace=namespace,
            body=service
        )
        
        return InfrastructureResource(
            resource_id=f"{namespace}/{name}",
            name=name,
            resource_type=InfrastructureType.LOAD_BALANCER,
            provider=CloudProvider.KUBERNETES,
            region="",
            config=config,
            status="creating",
            metadata={
                "namespace": namespace,
                "service_type": config.get("service_type", "ClusterIP"),
                "port": config.get("port", 80)
            }
        )
    
    async def _create_ingress(
        self,
        config: Dict[str, Any],
        tags: Dict[str, str] = None
    ) -> InfrastructureResource:
        """Create Kubernetes ingress"""
        
        name = config["name"]
        namespace = config.get("namespace", "default")
        
        labels = {"app": name}
        if tags:
            labels.update(tags)
        
        ingress = kubernetes.client.V1Ingress(
            metadata=kubernetes.client.V1ObjectMeta(
                name=name,
                namespace=namespace,
                labels=labels,
                annotations=config.get("annotations", {})
            ),
            spec=kubernetes.client.V1IngressSpec(
                rules=[
                    kubernetes.client.V1IngressRule(
                        host=config.get("host"),
                        http=kubernetes.client.V1HTTPIngressRuleValue(
                            paths=[
                                kubernetes.client.V1HTTPIngressPath(
                                    path=config.get("path", "/"),
                                    path_type="Prefix",
                                    backend=kubernetes.client.V1IngressBackend(
                                        service=kubernetes.client.V1IngressServiceBackend(
                                            name=config.get("service_name"),
                                            port=kubernetes.client.V1ServiceBackendPort(
                                                number=config.get("service_port", 80)
                                            )
                                        )
                                    )
                                )
                            ]
                        )
                    )
                ]
            )
        )
        
        response = self.networking_v1.create_namespaced_ingress(
            namespace=namespace,
            body=ingress
        )
        
        return InfrastructureResource(
            resource_id=f"{namespace}/{name}",
            name=name,
            resource_type=InfrastructureType.NETWORK,
            provider=CloudProvider.KUBERNETES,
            region="",
            config=config,
            status="creating",
            metadata={
                "namespace": namespace,
                "host": config.get("host"),
                "path": config.get("path", "/")
            }
        )
