services/code-generation/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
jinja2==3.1.2
sqlalchemy==2.0.23
httpx>=0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
anthropic==0.8.1
google-generativeai==0.3.2
databases[postgresql]==0.8.0
asyncpg==0.29.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
networkx==3.2.1
graphviz==0.20.1
python-graphviz==0.20.1

# Phase 2 ML and Analytics Dependencies
scikit-learn==1.3.2
scipy==1.11.4
tensorflow==2.15.0
torch==2.1.2
transformers==4.36.2
websockets==12.0
redis==5.0.1
pymongo==4.6.0
kafka-python==2.0.2

# Phase 2 Enterprise Integration Dependencies
boto3==1.34.0
azure-identity==1.15.0
azure-storage-blob==12.19.0
google-cloud-storage==2.10.0
paramiko==3.4.0
requests-oauthlib==1.3.1
aiohttp==3.9.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
